Do not expose Relay Environment
It's important to use the provided environment in the component instead of referencing the global one. It's not very obvious in simple applications but it's  very important later when you may have multiple environments. Better to do it properly from the very beginning. :)

I will update the @kiwicom/relay package.